requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.12.0
django-crispy-forms>=1.9.2,<1.10.0
requests>=2.24.0,<2.25.0
django-cleanup>=5.0.0
pdfminer3k==1.3.4
plotly==4.9.0